f32 -> f64

+++ wasm/src/util.rs
@@ -3,6 +3,6 @@ pub fn gaussian(x: i32, y: i32, sigma: f64) -> f64 {
         / (2.0 * std::f64::consts::PI * sigma * sigma)
 }
 
-pub fn map(x: f32, in_min: f32, in_max: f32, out_min: f32, out_max: f32) -> f32 {
+pub fn map(x: f64, in_min: f64, in_max: f64, out_min: f64, out_max: f64) -> f64 {
     (x - in_min) * (out_max - out_min) / (in_max - in_min) + out_min
 }
